+++ nequip/ase/nequip_calculator.py
@@ -4,6 +4,7 @@ import torch
 
 import ase.data
 from ase.calculators.calculator import Calculator, all_changes
+from ase.stress import full_3x3_to_voigt_6_stress
 
 from nequip.data import AtomicData, AtomicDataDict
 from nequip.data.transforms import TypeMapper
@@ -24,7 +25,7 @@ class NequIPCalculator(Calculator):
 
     """
 
-    implemented_properties = ["energy", "energies", "forces"]
+    implemented_properties = ["energy", "energies", "forces", "stress", "free_energy"]
 
     def __init__(
         self,
@@ -115,18 +116,18 @@ class NequIPCalculator(Calculator):
 
         # predict + extract data
         out = self.model(data)
-        forces = out[AtomicDataDict.FORCE_KEY].detach().cpu().numpy()
-        energy = (
-            out[AtomicDataDict.TOTAL_ENERGY_KEY].detach().cpu().numpy().reshape(tuple())
-        )
-
-        # store results
-        self.results = {
-            "energy": energy * self.energy_units_to_eV,
-            # force has units eng / len:
-            "forces": forces * (self.energy_units_to_eV / self.length_units_to_A),
-        }
-
+        self.results = {}
+        # only store results the model actually computed to avoid KeyErrors
+        if AtomicDataDict.TOTAL_ENERGY_KEY in out:
+            self.results["energy"] = self.energy_units_to_eV * (
+                out[AtomicDataDict.TOTAL_ENERGY_KEY]
+                .detach()
+                .cpu()
+                .numpy()
+                .reshape(tuple())
+            )
+            # "force consistant" energy
+            self.results["free_energy"] = self.results["energy"]
         if AtomicDataDict.PER_ATOM_ENERGY_KEY in out:
             self.results["energies"] = self.energy_units_to_eV * (
                 out[AtomicDataDict.PER_ATOM_ENERGY_KEY]
@@ -135,3 +136,16 @@ class NequIPCalculator(Calculator):
                 .cpu()
                 .numpy()
             )
+        if AtomicDataDict.FORCE_KEY in out:
+            # force has units eng / len:
+            self.results["forces"] = (
+                self.energy_units_to_eV / self.length_units_to_A
+            ) * out[AtomicDataDict.FORCE_KEY].detach().cpu().numpy()
+        if AtomicDataDict.STRESS_KEY in out:
+            stress = out[AtomicDataDict.STRESS_KEY].detach().cpu().numpy()
+            stress = stress.reshape(3, 3) * (
+                self.energy_units_to_eV / self.length_units_to_A**3
+            )
+            # ase wants voigt format
+            stress_voigt = full_3x3_to_voigt_6_stress(stress)
+            self.results["stress"] = stress_voigt