+++ nequip/nn/embedding/_one_hot.py
@@ -10,12 +10,15 @@ from .._graph_mixin import GraphModuleMixin
 
 @compile_mode("script")
 class OneHotAtomEncoding(GraphModuleMixin, torch.nn.Module):
+    """Copmute a one-hot floating point encoding of atoms' discrete atom types.
+
+    Args:
+        set_features: If ``True`` (default), ``node_features`` will be set in addition to ``node_attrs``.
+    """
+
     num_types: int
     set_features: bool
 
-    # TODO: use torch.unique?
-    # TODO: type annotation
-    # Docstrings
     def __init__(
         self,
         num_types: int,
@@ -33,7 +36,7 @@ class OneHotAtomEncoding(GraphModuleMixin, torch.nn.Module):
             ]
         self._init_irreps(irreps_in=irreps_in, irreps_out=irreps_out)
 
-    def forward(self, data: AtomicDataDict.Type):
+    def forward(self, data: AtomicDataDict.Type) -> AtomicDataDict.Type:
         type_numbers = data[AtomicDataDict.ATOM_TYPE_KEY].squeeze(-1)
         one_hot = torch.nn.functional.one_hot(
             type_numbers, num_classes=self.num_types