+++ nequip/scripts/benchmark.py
@@ -1,9 +1,8 @@
-import sys
 import argparse
 import textwrap
 import tempfile
-import contextlib
 import itertools
+import time
 
 import torch
 from torch.utils.benchmark import Timer, Measurement
@@ -12,10 +11,11 @@ from torch.utils.benchmark.utils.common import trim_sigfig, select_unit
 from e3nn.util.jit import script
 
 from nequip.utils import Config
-from nequip.data import AtomicData, dataset_from_config
+from nequip.data import AtomicData, AtomicDataDict, dataset_from_config
 from nequip.model import model_from_config
 from nequip.scripts.deploy import _compile_for_deploy
-from nequip.scripts.train import _set_global_options, default_config, check_code_version
+from nequip.scripts.train import default_config, check_code_version
+from nequip.utils._global_options import _set_global_options
 
 
 def main(args=None):
@@ -56,7 +56,6 @@ def main(args=None):
         default=1,
     )
 
-    # TODO: option to profile
     # TODO: option to show memory use
 
     # Parse the args
@@ -75,34 +74,67 @@ def main(args=None):
 
     # Load dataset to get something to benchmark on
     print("Loading dataset... ")
-    # Currently, pytorch_geometric prints some status messages to stdout while loading the dataset
-    # TODO: fix may come soon: https://github.com/rusty1s/pytorch_geometric/pull/2950
-    # Until it does, just redirect them.
-    with contextlib.redirect_stdout(sys.stderr):
-        dataset = dataset_from_config(config)
+    dataset_time = time.time()
+    dataset = dataset_from_config(config)
+    dataset_time = time.time() - dataset_time
+    print(f"    loading dataset took {dataset_time:.4f}s")
+    dataset_rng = torch.Generator()
+    dataset_rng.manual_seed(config.get("dataset_seed", config.get("seed", 12345)))
     datas = [
         AtomicData.to_AtomicDataDict(dataset[i].to(device))
-        for i in torch.randperm(len(dataset))[: args.n_data]
+        for i in torch.randperm(len(dataset), generator=dataset_rng)[: args.n_data]
     ]
     n_atom: int = len(datas[0]["pos"])
-    assert all(len(d["pos"]) == n_atom for d in datas)  # TODO handle the general case
-    # TODO: show some stats about datas
+    if not all(len(d["pos"]) == n_atom for d in datas):
+        raise NotImplementedError(
+            "nequip-benchmark does not currently handle benchmarking on data frames with variable number of atoms"
+        )
+    print(
+        f"    loaded dataset of size {len(dataset)} and sampled --n-data={args.n_data} frames"
+    )
+    # print some dataset information
+    print("    benchmark frames statistics:")
+    print(f"         number of atoms: {n_atom}")
+    print(f"         number of types: {dataset.type_mapper.num_types}")
+    print(
+        f"          avg. num edges: {sum(d[AtomicDataDict.EDGE_INDEX_KEY].shape[1] for d in datas) / len(datas)}"
+    )
+    avg_edges_per_atom = torch.mean(
+        torch.cat(
+            [
+                torch.bincount(
+                    d[AtomicDataDict.EDGE_INDEX_KEY][0],
+                    minlength=d[AtomicDataDict.POSITIONS_KEY].shape[0],
+                ).float()
+                for d in datas
+            ]
+        )
+    ).item()
+    print(f"         avg. neigh/atom: {avg_edges_per_atom}")
 
+    # cycle over the datas we loaded
     datas = itertools.cycle(datas)
 
+    # short circut
     if args.n == 0:
         print("Got -n 0, so quitting without running benchmark.")
         return
 
     # Load model:
-    print("Loading model... ")
+    print("Building model... ")
+    model_time = time.time()
     model = model_from_config(config, initialize=True, dataset=dataset)
+    model_time = time.time() - model_time
+    print(f"    building model took {model_time:.4f}s")
     print("Compile...")
     # "Deploy" it
     model.eval()
+    compile_time = time.time()
     model = script(model)
-
     model = _compile_for_deploy(model)
+    compile_time = time.time() - compile_time
+    print(f"    compilation took {compile_time:.4f}s")
+
     # save and reload to avoid bugs
     with tempfile.NamedTemporaryFile() as f:
         torch.jit.save(model, f.name)
@@ -138,8 +170,11 @@ def main(args=None):
                 p.step()
     else:
         print("Warmup...")
+        warmup_time = time.time()
         for _ in range(warmup):
             model(next(datas).copy())
+        warmup_time = time.time() - warmup_time
+        print(f"    {warmup} calls of warmup took {warmup_time:.4f}s")
 
         print("Starting...")
         # just time