+++ nequip/utils/test.py
@@ -27,6 +27,7 @@ def assert_permutation_equivariant(
     func: GraphModuleMixin,
     data_in: AtomicDataDict.Type,
     tolerance: Optional[float] = None,
+    raise_error: bool = True,
 ):
     r"""Test the permutation equivariance of ``func``.
 
@@ -39,7 +40,7 @@ def assert_permutation_equivariant(
         data_in: the example input data to test with
     """
     # Prevent pytest from showing this function in the traceback
-    # __tracebackhide__ = True
+    __tracebackhide__ = True
 
     if tolerance is None:
         atol = PERMUTATION_FLOAT_TOLERANCE[torch.get_default_dtype()]
@@ -56,11 +57,15 @@ def assert_permutation_equivariant(
     n_node: int = len(data_in[AtomicDataDict.POSITIONS_KEY])
     while True:
         node_perm = torch.randperm(n_node, device=device)
+        if n_node <= 1:
+            break  # otherwise inf loop
         if not torch.all(node_perm == torch.arange(n_node, device=device)):
             break
     n_edge: int = data_in[AtomicDataDict.EDGE_INDEX_KEY].shape[1]
     while True:
         edge_perm = torch.randperm(n_edge, device=device)
+        if n_edge <= 1:
+            break  # otherwise inf loop
         if not torch.all(edge_perm == torch.arange(n_edge, device=device)):
             break
     # ^ note that these permutations are maps from the "to" index to the "from" index
@@ -117,9 +122,14 @@ def assert_permutation_equivariant(
                     problems.append(
                         f"edge/node permutation invariance violated for field {k}; maximum componentwise error: {err:e}. (`{k}` was assumed to be invariant, should it have been marked as equivariant?)"
                     )
-    if len(problems) > 0:
-        raise AssertionError("\n".join(problems))
-    return
+    msg = "\n".join(problems)
+    if len(problems) == 0:
+        return
+    else:
+        if raise_error:
+            raise AssertionError(msg)
+        else:
+            return msg
 
 
 def assert_AtomicData_equivariant(
@@ -152,10 +162,16 @@ def assert_AtomicData_equivariant(
     if not isinstance(data_in, list):
         data_in = [data_in]
     data_in = [AtomicData.to_AtomicDataDict(d) for d in data_in]
+    device, dtype = (
+        data_in[0][AtomicDataDict.POSITIONS_KEY].device,
+        data_in[0][AtomicDataDict.POSITIONS_KEY].dtype,
+    )
 
     # == Test permutation of graph nodes ==
-    # TODO: since permutation is distinct, run only on one.
-    assert_permutation_equivariant(func, data_in[0], tolerance=permutation_tolerance)
+    # since permutation is discrete and should not be data dependent, run only on one frame.
+    permutation_problems = assert_permutation_equivariant(
+        func, data_in[0], tolerance=permutation_tolerance, raise_error=False
+    )
 
     # == Test rotation, parity, and translation using e3nn ==
     irreps_in = {k: None for k in AtomicDataDict.ALLOWED_KEYS}
@@ -178,21 +194,40 @@ def assert_AtomicData_equivariant(
             # must be this to actually rotate it
             irps[AtomicDataDict.CELL_KEY] = "3x1o"
 
+    stress_keys = (AtomicDataDict.STRESS_KEY, AtomicDataDict.VIRIAL_KEY)
+    for k in stress_keys:
+        irreps_in.pop(k, None)
+    if any(k in irreps_out for k in stress_keys):
+        from e3nn.io import CartesianTensor
+
+        stress_cart_tensor = CartesianTensor("ij=ji")  # stress is symmetric
+        stress_rtp = stress_cart_tensor.reduced_tensor_products().to(device, dtype)
+        # symmetric 3x3 cartesian tensor as irreps
+        for k in stress_keys:
+            irreps_out[k] = stress_cart_tensor
+
     def wrapper(*args):
         arg_dict = {k: v for k, v in zip(irreps_in, args)}
         # cell is a special case
-        if AtomicDataDict.CELL_KEY in arg_dict:
-            # unflatten
-            cell = arg_dict[AtomicDataDict.CELL_KEY]
-            assert cell.shape[-1] == 9
-            arg_dict[AtomicDataDict.CELL_KEY] = cell.reshape(cell.shape[:-1] + (3, 3))
+        for key in (AtomicDataDict.CELL_KEY,):
+            if key in arg_dict:
+                # unflatten
+                val = arg_dict[key]
+                assert val.shape[-1] == 9
+                arg_dict[key] = val.reshape(val.shape[:-1] + (3, 3))
         output = func(arg_dict)
         # cell is a special case
-        if AtomicDataDict.CELL_KEY in output:
-            # flatten
-            cell = arg_dict[AtomicDataDict.CELL_KEY]
-            assert cell.shape[-2:] == (3, 3)
-            arg_dict[AtomicDataDict.CELL_KEY] = cell.reshape(cell.shape[:-2] + (9,))
+        for key in (AtomicDataDict.CELL_KEY,):
+            if key in output:
+                # flatten
+                val = output[key]
+                assert val.shape[-2:] == (3, 3)
+                output[key] = val.reshape(val.shape[:-2] + (9,))
+        # stress is also a special case,
+        # we need it to be decomposed into irreps for equivar testing
+        for k in stress_keys:
+            if k in output:
+                output[k] = stress_cart_tensor.from_cartesian(output[k], rtp=stress_rtp)
         return [output[k] for k in irreps_out]
 
     # prepare input data
@@ -226,7 +261,9 @@ def assert_AtomicData_equivariant(
         problems = {k: v for k, v in errs.items() if v > o3_tolerance}
 
         def _describe(errors):
-            return "\n".join(
+            return (
+                permutation_problems + "\n" if permutation_problems is not None else ""
+            ) + "\n".join(
                 "(parity_k={:d}, did_translate={}) -> max error={:.3e}".format(
                     int(k[0]),
                     bool(k[1]),
@@ -235,7 +272,7 @@ def assert_AtomicData_equivariant(
                 for k, v in errors.items()
             )
 
-        if len(problems) > 0:
+        if len(problems) > 0 or permutation_problems is not None:
             raise AssertionError(
                 "Equivariance test failed for cases:" + _describe(problems)
             )
@@ -250,14 +287,16 @@ def assert_AtomicData_equivariant(
         problems = [e for e in all_errs if e[-1] > o3_tolerance]
 
         def _describe(errors):
-            return "\n".join(
+            return (
+                permutation_problems + "\n" if permutation_problems is not None else ""
+            ) + "\n".join(
                 "   (parity_k={:1d}, did_translate={:5}, field={:20}) -> max error={:.3e}".format(
                     int(k[0]), str(bool(k[1])), str(k[2]), float(k[3])
                 )
                 for k in errors
             )
 
-        if len(problems) > 0:
+        if len(problems) > 0 or permutation_problems is not None:
             raise AssertionError(
                 "Equivariance test failed for cases:\n" + _describe(problems)
             )
@@ -291,7 +330,7 @@ def set_irreps_debug(enabled: bool = False) -> None:
     from nequip.utils.torch_geometric import Data
 
     def pre_hook(mod: GraphModuleMixin, inp):
-        # __tracebackhide__ = True
+        __tracebackhide__ = True
         if not isinstance(mod, GraphModuleMixin):
             return
         mname = type(mod).__name__
@@ -327,7 +366,7 @@ def set_irreps_debug(enabled: bool = False) -> None:
     h1 = torch.nn.modules.module.register_module_forward_pre_hook(pre_hook)
 
     def post_hook(mod: GraphModuleMixin, _, out):
-        # __tracebackhide__ = True
+        __tracebackhide__ = True
         if not isinstance(mod, GraphModuleMixin):
             return
         mname = type(mod).__name__