+++ tests/unit/model/test_builder_utils.py
@@ -1,6 +1,7 @@
 import pytest
 import tempfile
 
+import numpy as np
 import torch
 
 import ase.io
@@ -8,12 +9,13 @@ import ase.io
 from nequip.data import AtomicDataDict, ASEDataset
 from nequip.data.transforms import TypeMapper
 
-from nequip.model.builder_utils import add_avg_num_neighbors
+from nequip.model.builder_utils import add_avg_num_neighbors, add_avg_num_atoms
 
 
 @pytest.mark.parametrize("r_max", [3.0, 2.0, 1.1])
 @pytest.mark.parametrize("subset", [False, True])
-def test_avg_num_neighbors(molecules, temp_data, r_max, subset):
+@pytest.mark.parametrize("to_test", ["neighbors", "atoms"])
+def test_avg_num(molecules, temp_data, r_max, subset, to_test):
     with tempfile.NamedTemporaryFile(suffix=".xyz") as fp:
         for atoms in molecules:
             # Reverse the atoms so the one without neighbors ends up at the end
@@ -36,34 +38,45 @@ def test_avg_num_neighbors(molecules, temp_data, r_max, subset):
             torch.randperm(len(nequip_dataset))[: len(nequip_dataset) // 2]
         )
 
+    func = {"neighbors": add_avg_num_neighbors, "atoms": add_avg_num_atoms}[to_test]
+
     # test basic options
-    annkey = "avg_num_neighbors"
+    annkey = f"avg_num_{to_test}"
     config = {annkey: 3}
-    add_avg_num_neighbors(config, initialize=False, dataset=None)
+    func(config, initialize=False, dataset=None)
     assert config[annkey] == 3.0  # nothing should happen
     assert isinstance(config[annkey], float)
 
     config = {annkey: 3}
     # dont need dataset if config isn't auto
-    add_avg_num_neighbors(config, initialize=False, dataset=None)
+    func(config, initialize=False, dataset=None)
     with pytest.raises(ValueError):
         # need if it is
         config = {annkey: "auto"}
-        add_avg_num_neighbors(config, initialize=True, dataset=None)
+        func(config, initialize=True, dataset=None)
 
     # compute dumb truth
-    num_neigh = []
-    for i in range(len(nequip_dataset)):
-        frame = nequip_dataset[i]
-        num_neigh.append(
-            torch.bincount(
-                frame[AtomicDataDict.EDGE_INDEX_KEY][0],
-                minlength=len(frame[AtomicDataDict.POSITIONS_KEY]),
-            ).float()
-        )
-    avg_num_neighbor_truth = torch.mean(torch.cat(num_neigh, dim=0))
+    if to_test == "neighbors":
+        num_neigh = []
+        for i in range(len(nequip_dataset)):
+            frame = nequip_dataset[i]
+            num_neigh.append(
+                torch.bincount(
+                    frame[AtomicDataDict.EDGE_INDEX_KEY][0],
+                    minlength=len(frame[AtomicDataDict.POSITIONS_KEY]),
+                ).float()
+            )
+        avg_num_truth = torch.mean(torch.cat(num_neigh, dim=0))
+    elif to_test == "atoms":
+        num_atoms = []
+        for i in range(len(nequip_dataset)):
+            frame = nequip_dataset[i]
+            num_atoms.append(frame.num_nodes)
+        avg_num_truth = np.mean(num_atoms)
+    else:
+        raise AssertionError
 
     # compare
     config = {annkey: "auto"}
-    add_avg_num_neighbors(config, initialize=True, dataset=nequip_dataset)
-    assert config[annkey] == avg_num_neighbor_truth
+    func(config, initialize=True, dataset=nequip_dataset)
+    assert config[annkey] == avg_num_truth